FreeBSD client: Updates from Mark Felder
git-svn-id: svn://svn.code.sf.net/p/xymon/code@7213 44351d6e-118b-4698-b696-ce33095ecaa4

+++ branches/4.3.13/client/freebsd-meminfo.c
@@ -23,7 +23,7 @@ int main(int argc, char *argv[])
 	unsigned long physmem;
 
 	int hw_pagesize[] = { CTL_HW, HW_PAGESIZE };
-	unsigned long pagesize;
+	int pagesize;
 
 	int vm_vmtotal[] = { CTL_VM, VM_METER };
 	struct vmtotal vmdata;
